Move the ContentRequest domain into its own module
UUIDInFilter is declared locally, as auth/viewsets does. OptionalPageNumberPagination
comes from kolibri.core.utils.pagination — content/api.py holds a duplicate of it
that is about to go.
