+++ kolibri/core/content/serializers.py
@@ -1,18 +1,11 @@
 from le_utils.constants import content_kinds
-from le_utils.constants.labels import learning_activities
 from rest_framework import serializers
 
 from kolibri.core.content.models import ChannelMetadata
-from kolibri.core.content.models import ContentDownloadRequest
 from kolibri.core.content.models import ContentNode
-from kolibri.core.content.models import ContentRemovalRequest
-from kolibri.core.content.models import ContentRequestReason
-from kolibri.core.content.models import FacilityUser
 from kolibri.core.content.models import File
 from kolibri.core.content.models import Language
-from kolibri.core.content.tasks import automatic_resource_import
 from kolibri.core.fields import create_timezonestamp
-from kolibri.core.serializers import KolibriModelSerializer
 
 
 class PublicChannelSerializer(serializers.ModelSerializer):
@@ -199,79 +192,3 @@ class ContentNodeGranularSerializer(serializers.ModelSerializer):
 
     def get_is_leaf(self, instance):
         return instance.kind != content_kinds.TOPIC
-
-
-class ContentDownloadRequestMetadataSerializer(serializers.Serializer):
-    title = serializers.CharField()
-    file_size = serializers.IntegerField()
-    learning_activities = serializers.ListField(
-        child=serializers.ChoiceField(learning_activities.choices)
-    )
-
-
-class ContentDownloadRequestSerializer(KolibriModelSerializer):
-    source_instance_id = serializers.UUIDField(
-        required=False, allow_null=True, write_only=True
-    )
-    metadata = serializers.JSONField()
-
-    class Meta:
-        model = ContentDownloadRequest
-        fields = (
-            "id",
-            "requested_at",
-            "reason",
-            "contentnode_id",
-            "metadata",
-            "status",
-            "facility",
-            "source_id",
-            "source_instance_id",
-        )
-        read_only_fields = (
-            "requested_at",
-            "reason",
-            "status",
-            "facility",
-            "source_id",
-        )
-
-    def validate_metadata(self, value):
-        # metadata is a JSONField, not this serializer: read returns the column
-        # verbatim, including the progress keys update_progress adds.
-        return ContentDownloadRequestMetadataSerializer().run_validation(value)
-
-    def create(self, validated_data):
-        # if there is an existing deletion request, delete the deletion request
-        if "request" in self.context and self.context["request"].user is not None:
-            user = self.context["request"].user
-        else:
-            raise serializers.ValidationError("User must be defined")
-
-        deletion_request = ContentRemovalRequest.objects.filter(
-            contentnode_id=validated_data["contentnode_id"],
-            source_id=user.id,
-            reason=ContentRequestReason.UserInitiated,
-            source_model=FacilityUser.morango_model_name,
-        )
-
-        deletion_request.delete()
-
-        existing_request = ContentDownloadRequest.objects.filter(
-            contentnode_id=validated_data["contentnode_id"],
-            source_id=user.id,
-            reason=ContentRequestReason.UserInitiated,
-            source_model=FacilityUser.morango_model_name,
-        ).first()
-
-        if existing_request:
-            return existing_request
-
-        content_request = ContentDownloadRequest.build_for_user(user)
-        content_request.metadata = validated_data["metadata"]
-        content_request.contentnode_id = validated_data["contentnode_id"]
-        content_request.source_instance_id = validated_data.get("source_instance_id")
-
-        content_request.save()
-        automatic_resource_import.enqueue_if_not_active()
-        return content_request
